Support loading a config for `dock` from ~/.config/dock
This config file is just a shell script that gets sourced, which
contains a bunch of variables that override a set of defaults. My reason
for this is because my work laptop and home laptop setups are a little
different (the work monitor being bigger and work laptop being closer to
my face).